+++ android/app/src/main/java/com/koleff/kare_android/ui/compose/navigation/Toolbar.kt
@@ -1,7 +1,6 @@
-package com.koleff.kare_android.ui.compose
+package com.koleff.kare_android.ui.compose.navigation
 
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material3.CenterAlignedTopAppBar
